[ADD/#21] 뷰잉파티 상세정보 조회 API 연동 세팅

+++ app/src/main/java/umc/everyones/lck/presentation/party/adapter/ViewingPartyRVA.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.RecyclerView
 import umc.everyones.lck.databinding.ItemViewingPartyBinding
 import umc.everyones.lck.domain.model.party.ViewingPartyItem
 
-class ViewingPartyRVA(val readViewingParty: (Int) -> Unit) :
+class ViewingPartyRVA(val readViewingParty: (Long) -> Unit) :
     ListAdapter<ViewingPartyItem, ViewingPartyRVA.ViewingPartyViewHolder>(DiffCallback()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewingPartyViewHolder {
